Create README.md
The "Web Scraping of List of Largest Companies in USA by Revenue from Wikipedia" project is a Jupyter Notebook-based data extraction tool that utilizes Python's BeautifulSoup library. The project's main objective is to scrape a table from a specific Wikipedia page, containing a comprehensive list of the largest companies in the USA ranked by revenue. The extracted data is organized into a structured format using Pandas DataFrame, and users are provided with the option to export the data to a spreadsheet for further analysis and exploration